TST: remove tests_tseries.py and distribute to other tests files
Author: Jeff Reback <jeff@reback.net>

Closes #13325 from jreback/reorg_tests and squashes the following commits:

c7e045e [Jeff Reback] TST: remove tests_tseries.py and distribute to other tests files
